# Faster cache key