--- a/modelator/Cargo.toml
+++ b/modelator/Cargo.toml
@@ -6,10 +6,11 @@ edition = "2018"
 
 [dependencies]
 clap = "3.0.0-beta.2"
+hex = "0.4.3"
 nom = "6.1.2"
 serde = { version = "1.0.123", features = ["derive"] }
 serde_json = "1.0.62"
-sha256 = "1.0.0"
+sha2 = "0.9.3"
 thiserror = "1.0.24"
 tracing = "0.1.25"
 tracing-subscriber = "0.2.16"