--- a/modelator/src/module/tlc/mod.rs
+++ b/modelator/src/module/tlc/mod.rs
@@ -25,7 +25,7 @@ impl Tlc {
 
         // load cache and check if the result is cached
         let mut cache = TlaTraceCache::new(options)?;
-        let cache_key = TlaTraceCache::key(&tla_file, &tla_config_file, options)?;
+        let cache_key = TlaTraceCache::key(&tla_file, &tla_config_file)?;
         if let Some(value) = cache.get(&cache_key)? {
             return Ok(value);
         }