The caching strategy introduced in #39 ended up not bringing any benefits to MBT in [ibc-rs](https://github.com/informalsystems/ibc-rs/): https://github.com/informalsystems/modelator/pull/39#issuecomment-806099190
This was due to the cost of computing the cache key (which was using `modelator::module::Apalache::parse`).

This PR introduces a new strategy for computing the cache key: read all TLA+ files in the directory where the TLA+ test file is and hash their content.

#### Potential issues:
- _cache is not used but should_: the user changes an unrelated TLA+ file and as a result the cached output is not used
- _cache is used but shouldn't_: the user changes some java code that changes the behaviour of some TLA+ module; in this case, the cache should be invalidated but it won't (arguably this cannot happen in the current version of `modelator` as we don't allow this kind of "imports")

#### Results for MBT in ibc-rs:
- not-cached:
  - before this PR: 32s
  - after this PR: 18s
- cached:
  - before this PR: 16s
  - after this PR: 1s