@andrey-kuprianov 2021-03-25
Cool, Vitor!

One thing I don't understand: where does this difference come from?

```
- not-cached:
    - before: 32s
    - after: 18s
```

These are the numbers from running on uncached results before and after the PR, right? Where does the speedup for uncached files result from?