@vitorenesduarte 2021-03-25
> These are the numbers from running on uncached results before and after the PR, right?

Exactly. (I'll update the PR description to make that clear)

> Where does the speedup for uncached files result from?

The overhead was in the computation of the cache key and that computation has to be done independently of whether the key is cached or not. I hope that makes sense.