@andrey-kuprianov 2021-03-25
Ah, I see now! The `before` is actually after the previous PR #39 merged, so with the slowdown from `apalache parse` included. yes, that makes sense. So, the diff from original is then

- before any caching: 20s
- after this PR (non-cached): 18s
- after this PR (cached): 1s

Looks good to me:)