@vitorenesduarte 2021-03-25
> So, the diff from original is then
> 
> * before any caching: 20s
> * after this PR (non-cached): 18s
> * after this PR (cached): 1s

Yes, that's a much better way to put it.
(BTW, the improvement for non-cached runs comes from https://github.com/informalsystems/modelator/commit/35ec551d8cf34d324a67719ca028335a652b689d.)